// src/app/todo/store/action-types.enum.ts
export enum TodoActionTypes {
  ADD_TODO = '[Todo] Add todo',
  ADD_TODO_SUCCESS = '[Todo] Add todo success',
  ADD_TODO_FAILURE = '[Todo] Add todo failure',

  REMOVE_TODO = '[Todo] Remove todo',
  REMOVE_TODO_SUCCESS = '[Todo] Remove todo success',
  REMOVE_TODO_FAILURE = '[Todo] Remove todo failure',

  GET_TODOS = '[Todo] Get todos',
  GET_TODOS_SUCCESS = '[Todo] Get todos success',
  GET_TODOS_FAILURE = '[Todo] Get todos failure',
}